pyproject.toml
[project]
name = "zander-ingestion-server"
version = "0.1.0"
description = "Lightweight bidirectional WebSocket broker for EEG data streams"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "fastapi[standard]>=0.122.0",
    "httpx>=0.28.1",
    "msgpack>=1.1.2",
    "mypy>=1.18.2",
    "prometheus-client>=0.23.1",
    "prometheus-fastapi-instrumentator>=7.1.0",
    "pydantic-settings>=2.12.0",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.1",
    "redis[hiredis]>=7.1.0",
    "ruff>=0.14.6",
    "sqlalchemy[asyncio]>=2.0.44",
    "uvicorn[standard]>=0.38.0",
    "websockets>=15.0.1",
]
